+++ src/TensorFlowNET.Core/Variables/gen_state_ops.py.cs
@@ -21,7 +21,7 @@ namespace Tensorflow
         /// <param name="container"></param>
         /// <param name="shared_name"></param>
         /// <returns></returns>
-        public static Tensor variable_v2(long[] shape, TF_DataType dtype, string name = null, string container = "", string shared_name = "")
+        public static Tensor variable_v2(int[] shape, TF_DataType dtype, string name = null, string container = "", string shared_name = "")
         {
             var _op = _op_def_lib._apply_op_helper("VariableV2", name: name, args: new { dtype, shape, container, shared_name });
